alfred: Add installation of the alfred-gpsd manpage
It is nice to have a manpage somewhere but the user is most likely more
interested in reading it from a familiar place. Installing it together with the
binary seems appropriate because then the user only has to call his/her prefered
manpage reader to get fast and easy access to information about everything the
alfred-gpsd binary provides.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Acked-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>