alfred: Handle EPERM on every sendto
The announcements are currently only sent by alfred instances in master mode.
Slave instances don't announce themself and thus don't detect the broken
socket. To also catch such problems, it is useful to handle the detection of
EPERM during the send in send_alfred_packet which is always called when a
link-local UDP packet is send over the network.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>