alfred: Check for changed interface properties
The interface may have been recreated or the mac address was changed. In this
situations socket has to be recreated to avoid proplems with the link-local
IPv6 address and the scope_id (interface id).

These properties are polled and checked every 60 seconds and the appropriate
measures are taken when an error/change occurs.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>