alfred-gpsd: Avoid underrun when reading from gpsd
The gpsd output reading function is ignoring \r characters. This is done by
moving the current position (cnt) one position back in the character buffer. It
is jumping to the -1 character (max number for size_t) when it was reading the
first character at position 0. This is not problematic when the cnt is
increased directly after it by 1. Overflows/underflows are defined for
*unsigned* types and thus it just jumps back to 0.

Unfortunatelly, it is trying to access the memory for another check before
increasing the position again. This check is done on memory outside of the
buffer and therefore invalid.

Instead doing two check after each other, it is in this situation better to do
both at once and just handle the current character.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>