alfred: Add support for multiple interfaces per master
It may required in some situations that all alfred master servers are
potentially connected to the same backbone network but cannot see each other
through the mesh. The data may still be exchanged and thus a master instance
must listen for requests on the mesh and synchronize data over the backbone
network.

The limitation that data can only be synchronized over one master hop and not
more still holds true. Also alfred servers in slave mode are not allowed to
operate on multiple interfaces. This is especially important when switching
from master to slave mode because this can fail now when multiple interfaces
are configured.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>
Signed-off-by: Simon Wunderlich <simon@open-mesh.com>