alfred: Remove bogus length check in debugfs_make_path
The length check never used the mesh_iface length in the size calculation.
Instead it used the length of the mountpoint and the format string. But the
length of the format string is not the length of the final string.

Instead remove this check and depend completely on the return value of
snprintf.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>