alfred: Switch netlink uapi header copyright to ISC
The batman_adv.h netlink header is used by userspace programs to use the
genl interface of batman-adv. These programs may not use the GPL-2 license.
The header should therefore state clearly that it can be used freely and
the license of the program which uses this header is not affected.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Acked-by: Matthias Schiffer <mschiffer@universe-factory.net>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>