alfred: Keep order of interfaces from commandline
The first (valid) interface mac address will be used as source mac for an
dataset. But alfred added the commandline parameters for a multi-interface
master in reverse order.

Keep the list in the same order as the commandline to fulfill the
user expectations when checking the source entries of datasets.

Signed-off-by: Sven Eckelmann <sven.eckelmann@open-mesh.com>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>