alfred: Only query tq of remote master in slave mode
The querying of the originator mac address and tq values of the orignator
address takes significant amount of time. It is therefore better to avoid
the TQ retrieval code when possible.

The TQ will not be used in master mode and it can therefore be skipped
together with the code which tries to find the new best server.

Signed-off-by: Sven Eckelmann <sven.eckelmann@openmesh.com>
Signed-off-by: Simon Wunderlich <simon.wunderlich@openmesh.com>