alfred: Switch packet.h license to MIT
The ISC license is considered as not recommended in "Linux kernel licensing
rules". It should only be used for existing code or for importing code from
a different project with that license.

Since we already use this license for the batman-adv related uapi headers,
we should also use this for our packet.h. This fixes at the same time the
SPDX license identifier which switched incorrectly to GPL-2.0 in commit
ea81a05080bd ("alfred: Add SPDX license identifier above copyright
header").

Signed-off-by: Sven Eckelmann <sven@narfation.org>