alfred: vis: Use rtnl to query list of hardifs of meshif
The normal way of network related programs to query the state of interfaces
is to use the rtnetlink. Most of these data can also be queried via sysfs
but it is better to use the same way for both retrieving the list of
interfaces and modifying the list of interfaces.

Also the sysfs files are deprecated and cause warnings when access:

  batman_adv: [Deprecated]: batadv-vis (pid 832) Use of sysfs file "mesh_iface".
  Use batadv genl family instead

In worst case, the file doesn't even exist when batman-adv was compiled
without sysfs support.

Reported-by: Linus Lüssing <linus.luessing@c0d3.blue>
Signed-off-by: Sven Eckelmann <sven@narfation.org>