alfred: Use primary/secondary to describe server roles
The terms master+slave are loaded and did not correctly describe the
relationship of the various levels of indirections in the alfred network
layout.

The components are now described as following:

* primary (servers):

  - announce themselves
  - synchronizing data between primary server
  - accepting data from secondary servers and clients
  - answering requests from secondary servers and clients

* secondary (servers)

  - forwarding data to primary server(s)
  - accepting data from clients
  - answering requests from clients

* [tertiary] clients

  - sending data to primary and secondary servers
  - requesting data from primary and secondary servers

The clients are only communicating over unix socket with the servers. They
can skip the secondaries and thus are better not described as tertiary in
the documentation. Their differentiation as "client" is sufficient to
differentiate them from primary/secondary servers.

Signed-off-by: Sven Eckelmann <sven@narfation.org>