alfred: Move next release to 2021
The first possible release day for the next release is in the last week of
december. But it can also happen that another rc8 will be added by Linus
Torvalds - and thus the release will be postponed to 2021. Just avoid this
uncertainty in the version number by directly going to 2021.0 and not doing
a feature release this year.

Signed-off-by: Sven Eckelmann <sven@narfation.org>